+++ tests/junit-functional/org/jgroups/tests/UNICAST_Test.java
@@ -61,8 +61,8 @@ public class UNICAST_Test {
                 {new UNICAST()}, {new UNICAST2()}
         };
 
-        ((UNICAST)retval[0][0]).setTimeout(new long[]{500,1000,2000,3000});
-        ((UNICAST2)retval[1][0]).setTimeout(new long[]{500,1000,2000,3000});
+        ((UNICAST)retval[0][0]).setTimeout(new int[]{500,1000,2000,3000});
+        ((UNICAST2)retval[1][0]).setTimeout(new int[]{500,1000,2000,3000});
         return retval;
     }
